SEP6-61: Finish view movie details plus general refactor
