+++ src/MovieManagement/Components/ImageComponent.razor.cs
@@ -1,0 +1,7 @@
+﻿namespace MovieManagement.Components;
+
+public partial class ImageComponent : ComponentBase
+{
+    [Parameter] 
+    public string ImageUrl { get; set; } = default!;
+}
