+++ src/MovieManagement/Components/ImageComponent.razor.css
