+++ src/MovieManagement/Components/MovieCard.razor
