+++ src/MovieManagement/Components/MovieCard.razor.css
