+++ src/MovieManagement/Components/MovieListView.razor.css
