+++ src/MovieManagement/GlobalUsings.cs
@@ -1,19 +1,25 @@
 global using Blazored.Modal;
-global using MovieManagement.Domain.Models;
-global using MovieManagement.ViewModels.MovieDetails;
-global using MovieManagement.Config;
-global using MovieManagement.Services;
-global using MovieManagement.ViewModels;
-global using MovieManagement.FunctionDtos;
-global using MovieManagement.TMDB.Api.Mappers;
-global using MovieManagement.TMDB.Api.Services;
-global using MovieManagement.TMDB.Api.Config;
 global using Microsoft.AspNetCore.Components;
-global using System.Text.Json;
-global using Microsoft.Extensions.Options;
-global using System.ComponentModel.DataAnnotations;
-global using MovieManagement.Models.Index;
-global using System.Security.Claims;
 global using Microsoft.AspNetCore.Components.Authorization;
+global using Microsoft.AspNetCore.Mvc;
+global using Microsoft.AspNetCore.Mvc.RazorPages;
+global using Microsoft.Extensions.Options;
 global using Microsoft.JSInterop;
 global using MovieManagement.Authentication;
+global using MovieManagement.Config;
+global using MovieManagement.Data;
+global using MovieManagement.Domain.Models;
+global using MovieManagement.FunctionDtos;
+global using MovieManagement.Services;
+global using MovieManagement.TMDB.Api.Config;
+global using MovieManagement.TMDB.Api.Mappers;
+global using MovieManagement.TMDB.Api.Services;
+global using MovieManagement.ViewModels;
+global using MovieManagement.ViewModels.User;
+global using MovieManagement.ViewModels.Index;
+global using MovieManagement.ViewModels.MovieDetails;
+global using System.ComponentModel.DataAnnotations;
+global using System.Diagnostics;
+global using System.Security.Claims;
+global using System.Text.Json;
+
