--- src/MovieManagement/Models/Index/MovieListViewModel.cs
@@ -1,7 +1,0 @@
-﻿namespace MovieManagement.Models.Index;
-
-public class MovieListViewModel
-{
-  public string Id { get; set; } = default!;
-  public string Name { get; set; } = default!;
-}
