--- src/MovieManagement/Models/Index/RatingViewModel.cs
@@ -1,7 +1,0 @@
-﻿namespace MovieManagement.Models.Index;
-
-public class RatingViewModel
-{
-  public string? Review { get; set; }
-  public int StarRating { get; set; } = default!;
-}
