+++ src/MovieManagement/Pages/MovieLists/InTheaterMovies.razor
@@ -1,3 +1,4 @@
 ﻿@page "/movies/in-theater"
+@using MovieManagement.Components
 
 <MovieListView ListType="ListType.InTheater" />
