+++ src/MovieManagement/Pages/MovieLists/Index.razor
@@ -1,4 +1,5 @@
 ﻿@page "/"
 @page "/movies/upcoming"
+@using MovieManagement.Components
 
 <MovieListView ListType="ListType.Upcoming"/>
