+++ src/MovieManagement/Pages/MovieLists/PopularMovies.razor
@@ -1,3 +1,4 @@
 ﻿@page "/movies/popular"
+@using MovieManagement.Components
 
 <MovieListView ListType="ListType.Popular" />
