+++ src/MovieManagement/Pages/MovieLists/TopRatedMovies.razor
@@ -1,3 +1,4 @@
 ﻿@page "/movies/top-rated"
+@using MovieManagement.Components
 
 <MovieListView ListType="ListType.TopRated" />
