--- src/MovieManagement/Shared/ImageComponent.razor.cs
@@ -1,6 +1,0 @@
-﻿namespace MovieManagement.Shared;
-
-public partial class ImageComponent : ComponentBase
-{
-   [Parameter] public string ImageUrl { get; set; } = default!;
-}
