+++ src/MovieManagement/ViewModels/Index/MovieListViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace MovieManagement.ViewModels.Index;
+
+public class MovieListViewModel
+{
+    public string Id { get; set; } = default!;
+    public string Name { get; set; } = default!;
+}
