+++ src/MovieManagement/ViewModels/Index/RatingViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace MovieManagement.ViewModels.Index;
+
+public class RatingViewModel
+{
+    public string? Review { get; set; }
+    public int StarRating { get; set; } = default!;
+}
